Move BlobStorageLogWriter to Interpreters
