+++ src/IO/WriteBufferFromS3.cpp
@@ -16,7 +16,7 @@
 #include <IO/S3Common.h>
 #include <IO/S3/Requests.h>
 #include <IO/S3/getObjectInfo.h>
-#include <IO/S3/BlobStorageLogWriter.h>
+#include <Interpreters/BlobStorageLogWriter.h>
 
 #include <utility>
